+++ nuvlaedge/agent/telemetry.py
@@ -371,13 +371,3 @@ class Telemetry(NuvlaEdgeCommon):
                           json.dumps(all_status), encoding='UTF-8')
 
         self.status.update(new_status)
-
-    def get_vpn_ip(self):
-        """ Discovers the NuvlaEdge VPN IP  """
-
-        if FILE_NAMES.VPN_IP_FILE.exists() and FILE_NAMES.VPN_IP_FILE.stat().st_size != 0:
-            with FILE_NAMES.VPN_IP_FILE.open('r') as vpn_file:
-                return vpn_file.read().splitlines()[0]
-        else:
-            logging.warning("Cannot infer the NuvlaEdge VPN IP!")
-            return None
